Feature/GetB2BUsersV2 implementation and tests moved to separate files
