+++ client/account/account.go
@@ -223,25 +223,6 @@ func (a *Account) GetUserByID(ctx context.Context, id int) (model.GetB2BUserByID
 	return resp, err
 }
 
-func (a *Account) GetB2BUsersV2(ctx context.Context, req model.GetB2BUsersV2Request) (model.GetB2BUsersV2Response, error) {
-	log.Trace("GetB2BUsersV2 called")
-
-	body, err := json.Marshal(req)
-	if err != nil {
-		return model.GetB2BUsersV2Response{}, errors.Wrap(err, "couldn't marshal input parameters")
-	}
-
-	r, err := a.cli.Post(ctx, model.URLB2BGetUsersV2, bytes.NewBuffer(body), nil)
-	if err != nil {
-		return model.GetB2BUsersV2Response{}, errors.Wrap(err, "can't get b2b users")
-	}
-
-	var resp model.GetB2BUsersV2Response
-	err = utils.UnmarshalAndCheckOk(&resp, r)
-
-	return resp, err
-}
-
 func getRefreshDuration(v int64) time.Duration {
 	expTime := time.Unix(v, 0).Add(-time.Minute * 1) // make trigger little-bit earlier than exp time
 	refreshDuration := expTime.UTC().Sub(time.Now())
